Carpet Cost Calculator
This was a simple Java program that basically incorporates a few classes
together in order to calculate a carpet cost. This was extremely simple
but going through the motions of writing code feels good. Also this was
the first time I've added a local repository to my github from the
command-line and without the help of an IDE.