Accidentally forgot to add in a Fish() class to my Inheritance project.
This class extends the Animal() class and simply stores the variables
and methods to move a fish tail. Calls on the Animal.move() method. I
added this class because it shows that we need to keep our parent
classes general. The objects of type Dog and Fish are different but both
can be represented by the Animal class (Parent class of both)

- Coronavirus Learning
- 05/24/2020