Polymorphism Problems
These are a few programs that I created that use a simple base parent
class and with a few child classes that override the parent class
methods and display examples of how polymorphism can be a good solution
to repeated operations. I've learned this topic before but going back
over some of the principles and actually learning about when to use this
technique was a nice refresh.

Tuesday May 26, Coronavirus learning