BurgerPoly Program Completed
This is a simple program to pull together a few key concepts that I
wanted to review such as Polymorphism and Inheritance. I built a basic
burger class and then went on to build specialty burgers that added to
this basic burger. I had to duplicate some code such as the
printBurgerReceipt() which I'd like to revisit and find a better way to
implement such a method. Also in the CowboyBurger class, I needed to
take away ability for an object of this class to use a few methods
defined in the parent class BaseBurger so I used @Override with empty
methods and I just feel like there has to be a better way.

Wednesday 27 2020, Coronavirus Learning