Array Learning and Challenge
This commit includes the start of an Array review module. There is a
simple program that goes over the basics of Java Arrays as well as an
ArrayChallenge1 class that inputs numbers from the keyboard into an
array, prints the array and then sorts the array in Descending order.
This was a very quick refresher on the Arrays Class and also I forgot
all about some of the quick built-in options for copying arrays such as
Arrays.copyOf(arrayToCopy, arrayToCopyLength). I might look into some
performance algorithms for array operations tonight if I get the chance.

Wed May 27th, Coronavirus Learning