Arrays refresh and a few coding challenges
This was a simple refresher on java array data structure. It was good to
go back over some of the fine details and to refresh some of the
built-in operations that help display arrays and also techniques to make
array algorithms better suited for outside changes. Also an interesting
idea I found was the placement of a Scanner object outside of a function
that allows us to use it in the class without having it created in
MAIN(). Kind of an interesting idea and I definitely could have used
this idea in some of my other coding problems.

Wed May 27th, Coronavirus Learning