Autoboxing and unboxing Refresher
This commit includes some code that refreshes my memory on autoboxing
techniques in java. I had forgot some of the shortcuts in java that make
for easier to read code and I'm glad that I went back over some of those
concepts. The next commit will be a small project that uses these
techniques in an actual program.

Wednesday June 3, Coronavirus Learning