Banking Application
This application simply goes over the key concepts of autoboxing and
unboxing (wrapper classes). This was an interesting topic because it
allowed me to fully understand why ArrayLists only store objects and not
primary data types. Also and intersting side topic that I picked up on
was limiting the doubling of computation within a single function.
I did this by storing the result of a computation in an object and then
just checked what I already had stored throughout the function without
needing to re-compute.

Tuesday June 30th, Coronavirus Learning