LinkedList Review
This was a simple review of LinkedLists and the basic functionalities
built-in to java. Reviewed Iterator objects and refreshed the idea that
an interator of type <String> needs to be initialized after creation or
else it won't point to the first item in a LinkedList<String> structure.
Now I am going to build my own small project to solidify my
understanding of LinkedLists.

July 6th 2020. Coronavirus Learning