Basic project based around implementing interfaces
I wanted to write this program so that I could better understand the
necessity for implementing interfaces. My program seemed to be better
adaptable to a wide range of use cases as well as be generic enough to
focus information in the classes themselves. The concept of keeping code
generic helped when I needed to write code that would operate with
multiple classes of the same interface.

07/30/2020 Coronavirus learning