Inner Classes Review
This was a refresh of the four basic types of inner classes (Local,
Anonymous, Static and non-static nested classes). I went ahead and
created OnClickListeners in a few different ways using the above basic
types of inner classes. Also this was a refresh on how to hide
implementation details that are a needed in a larger class but should
also be hidden away from the use cases of the larger class. In this
project the inner class "Gear" was needed in the "Gearbox" class but the
details of how "Gear" worked were not visible when using an object of
type Gearbox. Next I will build my own program to solidify my
understanding of Inner classes.

07/31/2020 Coronavirus Learning