Abstract Classes VS Interfaces
Today I went over the different use cases for Abstract classes and
interfaces and I worked on the structure that both allow or don't allow
when being use. I found it interesting looking into how an interface
might cause a great deal of problems later on in a program when what was
needed instead was the use of an abstract class. This topic was
enjoyable because I was able to focus on program structure again and
this has always been a favor topic of mine. Next, Java Generics

07/31/2020  Coronavirus learning