Simple project based around Interfaces
This project simply focused on how to structured programs that are
adaptable using interfaces. I was able to build multiple classes that
used the same interface with varying differences. I also modified the
interface to make the implementation more generic to open up the
possibilities of more use cases

Date 07/30/2020 CoronaVirus Learning