Package Demo Testing
In this current project I went about trying to understand how I could
write code and them import a jar file of the package into another
project and successfully run the code. I was able to figure this out
after learning the steps within the IDE to make this possible. It was
interesting to make a 'library' of sorts and then be able to import that
code into another project and run the code right away. It's extremely
exciting to feel like every hour I'm learning a great deal more about
both programming structure and development but also the tooling
requirements needed to be productive on my machine. I feel like I'm
starting to understand better how important it is to create a
comfortable development environment so that productivity can be
maximized.

08/04/2020 Coronavirus Learning