Scope Challenge Completed
This was a super cool challenge where I could only only really use 'X'
or 'x' as an identifier. For classes I could only use 'X' and for
objects of that class or any other variables I could only use 'x' as the
identifier. This was cool because it really forced me to think about
that specific scope relations of objects. I was forced to use "this.x"
and "innerClass.x" a few times and I even went as far as changing the
entire structure of the small program so that the Scanner object I used
to get data from the keyboard was called x. Overall a fun little
challenge that made me think which is good every once in a while.

08/04/2020 Coronavirus Learning