Java Collections Review/Study
In this program I worked with multiple classes within the Collections
interface so that I could see how easily they can be interchanged. I
worked on merge sorts, bubble sorts, and binary search algorithms as
well. This project as implemented inner classes which were nice to use
again as the more often I interact with this design structure, the
easier it will be for me to read and interpret code that uses it. Also
I've become increasingly curious about all of the methods within the
Collections library such as "swap", "reverse" and others that would be
nice to know for future applications. I really am starting to understand
why a common "tip" I hear is that a programmer should really work on
their toolbelt such as knowing the classes and libraries they use very
well so they don't spend time writing code that has already been
standardized and is faster to implement. Simply knowing a decent amount
of libraries and the common methods within them is a goal of mine in the
short term future.

08/06/2020 Coronavirus Learning