Map study
I added a non-related class to the Collections program in order to mess
around with the Map structure. When I built the Comparator object in the
Theatre class earlier today, I ran into an issue that was based on Maps
and I figured that I should take the time to go over maps while the
topic was still fresh. I simply worked with basic Map methods such as
remove, replace, put, etc. Now I'm going to build a small game with Maps
as the main topic of interest.

08/06/2020 Coronavirus Learning