Beginning of Adventure game with Maps
This is a remake of the classic game Adventure Cave from 1975. I wanted
to make this initial commit after I had laid down the very basic
structure of the game so if in the future I needed to backtrack to this
main point, I could. All I have built so far is a Location class so
store the locations of different exits and built up a small test suite
in the Main().

08/06/2020 Coronavirus Learning