Colossal Cave Adventure, before input splitting
This is the status of the game before I add in the code necessary to
allow players to type in directions with normal wordage instead of
simply typing 'N' for north, they can type "Run North" and the program
will still work. This is a rather large part of what made Colossal Cave
Adventure so important and popular in the 1970's so I want to get this
right. Also the previous ways of moving such as 'N' should still work
even after this next commit

08/06/2020 Coronavirus Learning