Completed the word processing and main refactor
This commit was loaded with changes (Might have made more changes than
I'd like with a single commit). To begin with I finished the basic "word
processing methods" by splitting the user input and then searching for
keywords in the string[]. Of course being able to use these libraries
made my job much easier than what Will Crowther had to do in the 70's
(hats off to him). I then went about refactoring the entire Main()
method in this project because it was taking on a create deal of
responsibility which needed to be placed in their own methods. All in
all, 5 methods came out of main() which just proves my earlier point of
why it needed to be refactored. I found this process very interesting as
I had to go through and decide which components should logically go with
others in separate methods. Now I need to figure out more ideas to
implement in this game while I have a little fire underneath me!

08/06/2020 Coronavirus Learning