Adventure Cave Finished
I have finished the this game/project and I am very happy with what I
accomplished in the game. I was able to work on immutable classes and
furthering my understanding of how objects state can be manipulated
outside of the class itself if careful consideration is not taken in
constructing the class. This of course helped me understand why
shallow/deep copies are so different both in practice and security of a
method. I have been reading "Clean Code" and I've been trying to
implement some of the ideas from that book into my code such as
constantly refactoring my code for ease of use and reading. I am going
to go over this Adventure game one more time looking for bugs/holes/ or
anything else that decreases the stability of the game. If I find
anything it will be provided in the next commit

08/08/2020 Coronavirus Learning