FIXED: NullPointerException Location Constructor
I fixed an bug in the constructor of the Location class where if a null
is passed in for the exits map, a NullPointerException would occur
during the initialization of the objects exits map. I fixed this bug by
testing the input parameters value and then giving clear paths to a bug
free creation of the object. It's always interesting passing in
unexpected values into methods and seeing what unexpected issues arise
that you would never have thought about up until this point. Expect the
unexpected I guess.

08/08/2020 Coronavirus Learning