Set and HashSet Work
I am building a project around using Sets as opposed to using lists and
the central theme will be the solar system. I am pretty excited about
this simply project because I've always enjoyed learning about space and
this will be a way to create some added interest in this project. I will
try to create objects for some of the defining objects in our solar
system such as planets, moons to which will orbit planets (maybe
astroids later on).

08/09/2020 Coronavirus Learning