Solar System Set project finished
I finished to Solar System Set project which was both fun but also
lacking due to my lack of understanding of building up graphics with
Java. I would really like to do a reboot of this project once I can
start building up a few graphics so that maybe I can actually program
and the watch the movement of the planetary bodies around their
respective orbit.

08/10/2020 Coronavirus Learning