Overriding equals() and hashCode
In the previous project I used an object from my Heavenly body class as
a value in a Set which means that I should override equals() and
hashCode(). I didn't realize this until I read some of the Java
documentation and so now I have gone back an implemented those changes.

08/10/2020 Coronavirus Learning