MyHashSet project creation
I was working on a leetcode problem and I ran into some issues with
tracking down where the program was creating an error so I rebuilt the
program in IntelliJ so that I could parse the situation data. This data
was comprised of the methods called in the test suite, the data used
in the method calls, the expected program output and the actual program
output. The test suite was comprised of 28 tests and the data was
getting far to long to parse myself just by scrolling about so I needed
to build a systematic way to not only replicate the issue but also see
which method call with which data created an error. I plan on publishing
a link to my github on leetcode once I finish the program so that others
can hopefully debug much faster.

08/11/2020 Coronavirus Learning