Testing subclass operations on equals() Override
One topic I came across was the idea that in particular cases, if we
allow a subclass to override the equals() method from the parent class,
we can run into issues with the basic logic of equals. Because subclass
objects are objects of the parent class but this is not the same case
vice versa, obj1 == obj2 is true but obj2 = obj1 is false. This of
course is a logical deviation from equals so when such operations are
needed, the equals() method that is overridden in the parent class
should be set to FINAL. We do not want the subclasses implementing their
own versions of equals() that only pertain to subclass objects.

08/11/2020 Coronavirus Learning