Gridpane Project
The current project covers the gridpane layout and how to configure the
children of this pane in the .fxml file. It is interesting to learn new
file structures such as .fxml compared to .xml files in android
development. After reading more into the MVC model
(model-view-controller) design method, I can't help but cringe at some
of my earlier Android projects where my application data & code was
mixed heavily with the UI code. It makes so much more sense to split UI
code and application code so that the program structure is separated and
clean. When I have a UI issues, I can go to the file containing only UI
code verses trying to parse UI code and application code. Really excited
to continue with the 7 other layouts in FX. I'm also looking forward to
learning how controllers link UI to application code and to structure
the data transfer that takes place. So much more to learn.

08/12/2020 Coronavirus Learning