Layout Study finished
I finished my basic study of javaFX layouts and now I will start
working with controllers and other features of this library. I am
excited to start to work with the structures of FX and to learn some of
the best practices. Also I wrote down the topic of creating my own
layout for future topics to look up.

08/15/2020 Coronavirus Learning