FX Controls Learning
This project focused on learning a few of the primary controls in the FX
libraries such as <Button> <RadioButton> <ToggleButton> <CheckBox>. I
also learned how to format some of the controls by placing multiple
controls into a child layout (VBox or HBox). Another cool concept that I
learned was how to import a .jar library of .gif icons into a JavaFX
project and then use the .gif icons in a button control. I always
wondered how this worked and it was nice to finally learn how to do it
myself through code instead of a SceneBuilder.

08/15/2020 Coronavirus Learning