Control Study Completed
I have gone through so many different controls in the FX library that I
feel like I have a decent grasp of many of the applications that there
are pre-built controls for. This is exactly the type of thing I wanted
to do early on so that I can start to get an idea of what I can
implement right away when I start building applications. Now I am going
to start studying events and event handlers so I can start to program
responses to UI interaction. I'm really excited to learn how to link the
UI to the processing code. Moving closer to my goals

08/15/2020 Coronavirus Learning