Thread Mania
I just learned about how to create multiple threads so that the UI
thread is not bogged down by slow processes. I of course had to make a
new Runnable object which I simply used a Thread.sleep(10000) operation
to handle like a long process. What was interesting was the idea that I
had to use Platform.runLater() with a new runnable object within this
other runnable object in order to update a component in the UI. I
believe the reason I have to do this is so that the UI stays thread
safe. We wouldn't want one thread to interact with another threads
actions while they are running.

08/15/2020 Coding Day #13