TodoList Application Foundation
I am creating a basic to-do list application just so that I can put what
I've studied to the test. This commit is after I've created the
foundation of the application such as getting packages in order,
creating a base item to be used in the to-do-list. I've made sure that
the application runs at this point and all of the basic start-up program
structure is set. I will continually add commits along the way as I add
features. Maybe I can start working with different git branches so I can
have a production branch along with some other branches as I develop
features in the application. Who knows, the sky is the limit.

08/15/2020 Coding Day# 15