Extra files from yesterdays Commits
These are extra files that were not added with yesterdays commit for
reasons I don't understand yet. When I would try to add them
individually they would not be staged for a commit. It was only after
using "git commit -a" that they actually showed up in the staging area
for this commit.

08/16/2020 Coding Day #14