Before Singleton Addition
I am at the stage in my program where I have everything working as I'd
want with mock data but now I need to switch the program to actually
store and then load data into the application. I will start with
creating a singleton class to store data and then I will build off of
that design. I wanted to make this commit before moving onto this larger
structural change so that if I want to restart the new build, I can
start with this commit.

08/16/2020 Coding Day #14