TodoData created
I have created a singleton class called todoData which will be used to
load & write application data (list items) to a file. At some point I
might convert the file data type to a commonly formatted version such as
JSON or CSV but for right now I'm just storing the data as strings
separated by a TAB on individual lines. One possible error I can see
with this is if a single entry is overly large and the line of the file
overflows into the next and then the data is shuffled. I will have to
test this as I just now thought of it as I write this commit.

08/17/2020 Day #15