Loader Complete, Dialog Controller created,
I have completed the data loader which is called in the Main.java class
in the init() method. Also the data writer is also completed and is
called in Main.java through the stop() method. I have created a dialog
popup that is accessed through the MenuBar in the mainWindow.fxml. I
created a corresponding Dialog controller class which processes user
input and updates our TodoData singleton data class.

08/17/2020 Day #15