FXCollections transition & textfield formatting
I changed some of the Collections objects in this program from Java
util. collection objects to an FX version. I have read a little about
how FX Collection structures are optimized for UI operations so I
figured I would make the switch while the program is still relatively
small. I also went about formatting the textfield to wrap text as when I
tried to add a 3k word details section to a new note, not only was the
input textfield hard to use without wrapping the text but the mainWindow
textfield was hard to use without the wrap text property set to true.

08/18/2020 Day #16