Cell Factory created, item colored based on data
I created a cell factory that colors the UI text of each item based on
when the deadline data is for each particular item. This allows for the
quick scanning of the list to figure out which item to prioritize. I
could also sort the list based on this same principle and change how
they are displayed (possible future change). I am now about to create a
deleting mechanism so that when a user goes to delete an item, a view is
displayed asking for validation of the action before the item is
deleted.

08/18/2020 Day#16