Delete Option added (Right Click ContextMenu)
I added a right-Click context menu delete option so that a user can
delete an item. I will also add a delete button option that will delete
the currenly selected/highlighted option as well as a keystroke listener
to delete the currently highlighted/selected list item when the DELETE
key is pressed. I also learned a little bit about Lambda functions which
I used in the Controller.java file to link the ContextMenu to each cell
created in the cell factory.

08/18/2020 Day#16