Basic Todo Application Complete
I finished much of the basic features for this application and so I
wanted to call it quits for now and make a commit when everything was
working as intended. I have so much more that I would like to change
with this application but I will wait until I've learned more of other
topics such as data structures and deeper Java subjects. This may be a
project I just pick up and work on every once in a while. I thought it
would be awesome to make a copy of the Apple Notes Application and see
how close I could get it to appearing the same.

08/16/2020 Day #16