Fri Aug 20 Days work
Today I worked on a multitude of topics and projects such as the Java
Memory Model, a simple contract app, and a refresh on SceneBuilder. I
really thought the material on the Java Memory Model was extremely
interesting and in finding a good resource on that topic, I also found a
great tutorials resource for java in general. Very excited to keep on
this journey.

08/20/2020 Day# 18