Adding file loader and saving methods.
I split the data into it's own package and now everything to do with the
Application data is together. I also wrote loader/saving methods for the
data. I will take the simple approach of saving data directly in the
application verses using external data base architecture. I think that
this will be the best approach for this application.

08/21/20 Day #19