Oracle Documentation Crawl
I am currently crawling my way through the oracle Java documentation as
I want to improve my ability to find resources that I need quickly. I've
already found a great deal of information in the documentation and it
has been a good experience so far. I am going to continue with this
study for a few hours today before switching to application building.

08/26/2020 Day 21