Small Nested Class Program
In this program I simply tested out a few situations related to Nested
Classes and how they related to the outer class they are in. I wanted to
nail down the scope differences between inner classes and static nested
classes. I also found it useful to work on creating objects of inner
classes and the minor differences in the syntax for the two

08/26/2020 Day 21