Starting my Java Study again
I had a fun time away from Java for a few days but now it is time to
continue my review of the language and the newest features of this
release. I really enjoyed game development in Python and I have access
to a free Game Development book using Java so maybe I take a peek at
this in a few days (hehe).

09/09/2020