Edit Feature Added
The edit feature was added to the contact application. When you select
a contact and then click "Edit..". This will prompt a dialog box with
the contacts data present to which you can modify and then save.
- Now I am moving onto a "Delete.." button that will delete a contact
only after an alert has asked if they want to actually delete
information.

09/25/2020