Refactoring and change to "Notes" property
I refactored "Controller.java" so that I could limit code duplication
when it came to producing INFORMATION Alert objects. I created a private
method called createInformationAlert which could be implemented in a variety of
ways in other methods.
For my second modification I made the "notes" property for a contact
basically optional by inserting a space in the XML file if no user input
string was provided for this property.

09/25/2020