ERROR PATCH: updateContact & further refactoring
I refactored how I handle the results of dialogs in the application such
that when an attempt to edit a contact and leave a required field empty
will cause an Alert (I am going to change this from an Alert to a dialog
display change in the next commit). I also added a strip() call to
"notes" in the case where the field was left empty and I needed to clear
the placeholder " " space when edited in the future to hold a value.

09/25/2020