Error fixed in "Add.." functionality
An error occurred when a user tried to make a new contact omitting
required information (first name, last name, phone number) to which an
Alert would be prompted but due to implementation issues, a mistake in
both the contacts.xml file and overal program logic would occur. With a
corrupted contacts.xml file, the program would not reopen without fixing
the file manually. I have cleared up this issue and now this
functionality should work as expected with all input types and
occurances.
- Now I need to move onto the other methods that will have a similar
error. Future commits will describe the issues and the fixes

09/25/2020