Error fixed in "Edit.." Feature
I fixed the issue described in the last commit but for this
functionality where I needed to handle the case where one might edit a
contact and leave out information when editing the contact information.
I fixed this issue and now I think I will try to refactor my code and
possible change where I validate a a successful contact creation
attempt. Right now I am handling all of this "checking" in the
Controller.java file but I almost think that I should transfer much of
the validation code to methods in the ContactController.java file.

09/25/2020