Alerts Removed, Hidden Label implemented
I moved altered an issue/error notification for the use from an Alert
onscreen to a slight modification to the fxml file that displays a
dialog for both the Add.. and Edit.. features. I have a hidden Label at
the bottom of the dialog that will appear if required fields are left
blank. This makes the entire process seem less clunky.

09/26/2020