Basic Application start
So far in the applicaiton I have set up a foundation with my main fxml
file, Controller java file and a css file. I have started to arrange
items in the applicaiton and do some default styling just to get a
better idea of how the layout will work and what I'm actually shooting
for. I've added the default buttons that will be the main naviagtion for
the application. I am going to finish a few css attributes and then move
onto linking actions and starting to get individual scenes built.

09/26/2020