Card class & Hover finish
I finished hover functionality so that when the mouse moves over buttons
they will highlight and go back to default when the mouse leaves their
boarders. I have not built in the code that keeps whatever current
page/scene the user is in highlighted on the left hand side but that
will come soon once I actually start building different scenes for each
button. I also started to work on the Card.class that will store the
most basic form of data for a review card such as frontside info and
backside info along with dateCreated and a few other fields. Not sure if
I'm leaving out important fields so I am going to give this topic a
little more thought before I dive into building a data model that
loads/writes card date to an internal xml file.
- Overall I'm just trying to have fun with this program and just throw
things against the wall and make a ton of errors while trying to see
what sticks.

09/26/2020