Nightly Commit
I am cashing in on the project for the day. I worked really hard on the
project while also reading a couple hundred pages of some other material
which made for a productive day. I am about to study data structures and
algorithms for a few hours tonight and maybe crack into my Operating
Systems material for a while as well. I will pick up on the project
tomorrow.

09/26/2020